Move resources to trview.app
Nothing works any more because the resources aren't linked yet.
#715
