+++ trview.app/Resources/DefaultFonts.h
