+++ trview.app/Resources/DefaultShaders.h
