+++ trview.app/Resources/DefaultTextures.h
