+++ trview.app/Resources/Files/arial7.bin
