+++ trview.app/Resources/Files/numericdown.png
