+++ trview.app/Resources/Files/numericup.png
