+++ trview.app/Resources/Files/settings.png
