+++ trview.app/Resources/Files/small.ico
