+++ trview.app/Resources/Files/texturelist.txt
