+++ trview.app/Resources/Files/tomb1.png
