+++ trview.app/Resources/Files/tomb2.png
