+++ trview.app/Resources/Files/tomb3.png
