+++ trview.app/Resources/Files/tomb4.png
