+++ trview.app/Resources/Files/tomb5.png
