+++ trview.app/Resources/Files/trview.ico
