+++ trview.app/Resources/Files/type_names.txt
