+++ trview.app/Resources/Files/unknown.png
