+++ trview.app/Resources/ResourceHelper.cpp
