+++ trview.app/Resources/ResourceHelper.h
