+++ trview.app/Resources/targetver.h
