+++ trview/stdafx.h
@@ -5,7 +5,7 @@
 
 #pragma once
 
-#include "targetver.h"
+#include <trview.app/Resources/targetver.h>
 
 #define NOMINMAX
 #define WIN32_LEAN_AND_MEAN
@@ -25,16 +25,9 @@
 #include <iomanip>
 #include <memory>
 
-#include <external/DirectXTK/Inc/WICTextureLoader.h>
 #include <directxmath.h>
 #include <commdlg.h>
 #include <shellapi.h>
 #include <Shlwapi.h>
 
-#include <trview.graphics/FontFactory.h>
-#include <trview.graphics/Device.h>
-#include <trview.graphics/IShaderStorage.h>
-#include <trview.graphics/VertexShader.h>
-#include <trview.graphics/PixelShader.h>
-#include <trview.app/Graphics/ITextureStorage.h>
 #include <trview.common/Strings.h>
