+++ trview/trview.vcxproj.filters
@@ -1,44 +1,16 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Filter Include="Graphics">
-      <UniqueIdentifier>{94ba7176-10d8-4cf9-b2f2-e6a6d554e70a}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Win">
       <UniqueIdentifier>{19537d19-9f7b-4110-a120-984dd4b941de}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Win\Resource Files">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Resources">
-      <UniqueIdentifier>{bb6233b6-2218-4c3e-8934-1ef700a66256}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="resource.h">
-      <Filter>Win</Filter>
-    </ClInclude>
     <ClInclude Include="stdafx.h">
-      <Filter>Win</Filter>
-    </ClInclude>
-    <ClInclude Include="targetver.h">
       <Filter>Win</Filter>
     </ClInclude>
     <ClInclude Include="trview.h">
       <Filter>Win</Filter>
-    </ClInclude>
-    <ClInclude Include="DefaultShaders.h">
-      <Filter>Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="DefaultTextures.h">
-      <Filter>Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="ResourceHelper.h">
-      <Filter>Graphics</Filter>
-    </ClInclude>
-    <ClInclude Include="DefaultFonts.h">
-      <Filter>Graphics</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -48,91 +20,5 @@
     <ClCompile Include="trview.cpp">
       <Filter>Win</Filter>
     </ClCompile>
-    <ClCompile Include="DefaultShaders.cpp">
-      <Filter>Graphics</Filter>
-    </ClCompile>
-    <ClCompile Include="DefaultTextures.cpp">
-      <Filter>Graphics</Filter>
-    </ClCompile>
-    <ClCompile Include="ResourceHelper.cpp">
-      <Filter>Graphics</Filter>
-    </ClCompile>
-    <ClCompile Include="DefaultFonts.cpp">
-      <Filter>Graphics</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ResourceCompile Include="trview.rc">
-      <Filter>Win\Resource Files</Filter>
-    </ResourceCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <Image Include="small.ico">
-      <Filter>Win\Resource Files</Filter>
-    </Image>
-    <Image Include="trview.ico">
-      <Filter>Win\Resource Files</Filter>
-    </Image>
-    <Image Include="resources\tomb3.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\unknown.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\tomb2.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\tomb1.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\tomb4.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\tomb5.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\numericdown.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\numericup.png">
-      <Filter>Resources</Filter>
-    </Image>
-    <Image Include="resources\settings.png">
-      <Filter>Resources</Filter>
-    </Image>
-  </ItemGroup>
-  <ItemGroup>
-    <Text Include="fontlist.txt">
-      <Filter>Win\Resource Files</Filter>
-    </Text>
-    <Text Include="resources\type_names.txt">
-      <Filter>Win\Resource Files</Filter>
-    </Text>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="texturelist.txt">
-      <Filter>Win\Resource Files</Filter>
-    </None>
-    <None Include="resources\level_pixel_shader.cso">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\level_vertex_shader.cso">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\ui_pixel_shader.cso">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\ui_vertex_shader.cso">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\arial8.bin">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\selection_pixel_shader.cso">
-      <Filter>Resources</Filter>
-    </None>
-    <None Include="resources\arial7.bin">
-      <Filter>Resources</Filter>
-    </None>
   </ItemGroup>
 </Project>
